disable memsan for now

+++ src/include/test.h
@@ -69,12 +69,12 @@ extern TestEntry tests[];
 	}
 
 #if MEMSAN == 1
-#define ASSERT_BYTES(v)                                              \
-	do {                                                         \
-		if (get_allocated_bytes() != v)                      \
-			printf("expected b=%d, found bytes=%d\n", v, \
-			       get_allocated_bytes());               \
-		ASSERT_EQ(get_allocated_bytes(), v, "ASSERT_BYTES"); \
+#define ASSERT_BYTES(v)                                                      \
+	do {                                                                 \
+		if (get_allocated_bytes() != v)                              \
+			printf("expected b=%ld, found bytes=%ld\n", (long)v, \
+			       (long)get_allocated_bytes());                 \
+		ASSERT_EQ(get_allocated_bytes(), v, "ASSERT_BYTES");         \
 	} while (0);
 #define ASSERT_NOT_BYTES_0() ASSERT(get_allocated_bytes(), "ASSERT_NOT_BYTES")
 #else
